Remove un-necessary doc override that we used to maintain for polyglot purpose
