+++ src/main/asciidoc/buffers.adoc
@@ -35,7 +35,12 @@ Create a buffer from a String: The String will be encoded using the specified en
 {@link examples.BufferExamples#example3}
 ----
 
-include::override/buffer_from_bytes.adoc[]
+Create a buffer from a byte[]
+
+[source,java]
+----
+{@link examples.BufferExamples#example4}
+----
 
 Create a buffer with an initial size hint. If you know your buffer will have a certain amount of data written to it
 you can create the buffer and specify this size. This makes the buffer initially allocate that much memory and is
