--- src/main/asciidoc/override/dependencies.adoc
@@ -1,22 +1,0 @@
-If you are using Maven or Gradle, add the following dependency to the _dependencies_ section of your
-project descriptor to access the Vert.x Core API:
-
-* Maven (in your `pom.xml`):
-
-[source,xml,subs="+attributes"]
-----
-<dependency>
-  <groupId>io.vertx</groupId>
-  <artifactId>vertx-core</artifactId>
-  <version>${maven.version}</version>
-</dependency>
-----
-
-* Gradle (in your `build.gradle` file):
-
-[source,groovy,subs="+attributes"]
-----
-dependencies {
-  compile 'io.vertx:vertx-core:${maven.version}'
-}
-----
