--- src/main/asciidoc/override/eventbus_headers.adoc
@@ -1,9 +1,0 @@
-==== Setting headers on messages
-
-Messages sent over the event bus can also contain headers. This can be specified by providing a
-{@link io.vertx.core.eventbus.DeliveryOptions} when sending or publishing:
-
-[source,$lang]
-----
-{@link docoverride.eventbus.Examples#headers(io.vertx.core.eventbus.EventBus)}
-----
