--- src/main/java/docoverride/buffer/Examples.java
@@ -1,27 +1,0 @@
-/*
- * Copyright (c) 2011-2019 Contributors to the Eclipse Foundation
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
- * which is available at https://www.apache.org/licenses/LICENSE-2.0.
- *
- * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
- */
-
-package docoverride.buffer;
-
-import io.vertx.core.buffer.Buffer;
-import io.vertx.docgen.Source;
-
-/**
- * @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
- */
-@Source
-public class Examples {
-
-  public void example4() {
-    byte[] bytes = new byte[] {1, 3, 5};
-    Buffer buff = Buffer.buffer(bytes);
-  }
-}
