+++ src/main/java/examples/BufferExamples.java
@@ -27,6 +27,11 @@ public class BufferExamples {
     Buffer buff = Buffer.buffer("some string");
   }
 
+  public void example4() {
+    byte[] bytes = new byte[] {1, 3, 5};
+    Buffer buff = Buffer.buffer(bytes);
+  }
+
   public void example3() {
     Buffer buff = Buffer.buffer("some string", "UTF-16");
   }
